fix: change position of shareable flag in catalog
